+++ app/src/main/AndroidManifest.xml
@@ -30,7 +30,7 @@
             android:exported="false"
             android:launchMode="singleTask" />
         <activity
-            android:name=".MainActivity"
+            android:name=".user.MainActivity"
             android:exported="false"
             android:launchMode="singleTask" />
         <activity
