+++ app/src/main/res/layout/fragment_user.xml
@@ -6,7 +6,7 @@
 
         <variable
             name="userViewModel"
-            type="com.nokopi.marketregistersystem.UserViewModel" />
+            type="com.nokopi.marketregistersystem.user.UserViewModel" />
 
     </data>
